psc: fix pm module init

+++ libraries/libstratosphere/include/stratosphere/psc/psc_pm_module_id.hpp
@@ -19,7 +19,7 @@
 
 namespace ams::psc {
 
-    enum PmModuleId : u16 {
+    enum PmModuleId : u32 {
         PmModuleId_Usb           = 4,
         PmModuleId_Ethernet      = 5,
         PmModuleId_Fgm           = 6,
